requirements.txt
pandas
pyarrow
python-dotenv
psycopg2-binary
duckdb
